requirements.txt
streamlit
google-cloud-aiplatform[preview]
beautifulsoup4
requests
python-docx
weasyprint